+++ Core/Executor/ContentTypeGroupManager.php
@@ -144,7 +144,11 @@ class ContentTypeGroupManager extends RepositoryExecutor implements MigrationGen
                     throw new \InvalidArgumentException('Content Type Group Manager does not support setting references for attribute ' . $reference['attribute']);
             }
 
-            $this->referenceResolver->addReference($reference['identifier'], $value);
+            $overwrite = false;
+            if (isset($reference['overwrite'])) {
+                $overwrite = $reference['overwrite'];
+            }
+            $this->referenceResolver->addReference($reference['identifier'], $value, $overwrite);
         }
 
         return true;