+++ Core/Executor/ObjectStateGroupManager.php
@@ -163,7 +163,11 @@ class ObjectStateGroupManager extends RepositoryExecutor implements MigrationGen
                     throw new \InvalidArgumentException('Object State Group Manager does not support setting references for attribute ' . $reference['attribute']);
             }
 
-            $this->referenceResolver->addReference($reference['identifier'], $value);
+            $overwrite = false;
+            if (isset($reference['overwrite'])) {
+                $overwrite = $reference['overwrite'];
+            }
+            $this->referenceResolver->addReference($reference['identifier'], $value, $overwrite);
         }
 
         return true;