+++ Core/Executor/SectionManager.php
@@ -149,7 +149,11 @@ class SectionManager extends RepositoryExecutor implements MigrationGeneratorInt
                     throw new \InvalidArgumentException('Section Manager does not support setting references for attribute ' . $reference['attribute']);
             }
 
-            $this->referenceResolver->addReference($reference['identifier'], $value);
+            $overwrite = false;
+            if (isset($reference['overwrite'])) {
+                $overwrite = $reference['overwrite'];
+            }
+            $this->referenceResolver->addReference($reference['identifier'], $value, $overwrite);
         }
 
         return true;