+++ Core/Executor/UserManager.php
@@ -247,7 +247,11 @@ class UserManager extends RepositoryExecutor
                     throw new \InvalidArgumentException('User Manager does not support setting references for attribute ' . $reference['attribute']);
             }
 
-            $this->referenceResolver->addReference($reference['identifier'], $value);
+            $overwrite = false;
+            if (isset($reference['overwrite'])) {
+                $overwrite = $reference['overwrite'];
+            }
+            $this->referenceResolver->addReference($reference['identifier'], $value, $overwrite);
         }
 
         return true;